+++ sentry/src/test/java/io/sentry/NoOpMemoryCollectorTest.kt
@@ -1,0 +1,13 @@
+package io.sentry
+
+import kotlin.test.Test
+import kotlin.test.assertNull
+
+class NoOpMemoryCollectorTest {
+    private var collector = NoOpMemoryCollector.getInstance()
+
+    @Test
+    fun `collect returns null`() {
+        assertNull(collector.collect())
+    }
+}