+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Collect memory usage in transactions ([#2445](https://github.com/getsentry/sentry-java/pull/2445))
+
 ### Fixes
 
 - Don't override sdk name with Timber ([#2450](https://github.com/getsentry/sentry-java/pull/2450))